backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
authlib==1.3.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
email-validator==2.1.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.12
alembic==1.14.0
bcrypt==4.2.1
python-multipart==0.0.6